changed template values to long long

+++ src/g2cdegrib2.c
@@ -46,7 +46,7 @@ extern G2C_CODE_TABLE_T *g2c_table;
  * @author Ed Hartnett @date Sep 28, 2022
  */
 int
-g2c_get_datetime(int ipdtn, int *ipdtmpl, short year, unsigned char month, unsigned char day,
+g2c_get_datetime(int ipdtn, long long int *ipdtmpl, short year, unsigned char month, unsigned char day,
                  unsigned char hour, unsigned char minute, unsigned char second, char *tabbrev)
 {
     int iutpos, iutpos2, iunit, iunit2;
@@ -275,7 +275,7 @@ format_level(char *cbuf, int ival, int scale)
  * @author Ed Hartnett @date Sep 28, 2022
  */
 int
-g2c_get_level_desc(int ipdtn, int *ipdtmpl, char *level_desc)
+g2c_get_level_desc(int ipdtn, long long int *ipdtmpl, char *level_desc)
 {
     int ipos;
     int ret;
@@ -681,8 +681,10 @@ g2c_degrib2(int g2cid, const char *fileout)
             else
                 fprintf(f, "  Num. of Data Points =  %d     NO BIT-MAP \n", sec5_info->num_data_points);
 	    fprintf(f, "  DRS TEMPLATE 5. %d : ", sec5_info->data_def);
+            /* Cast to int here. That's because degrib2.F90 shows
+             * these as signed ints, even if they are unsigned. */
             for (t = 0; t < sec5->template_len; t++)
-                fprintf(f, " %lld", sec5->template[t]);
+                fprintf(f, " %d", (int)sec5->template[t]);
             fprintf(f, "\n");
 	    fprintf(f, "  Data Values:\n");
 	    fprintf(f, "  Num. of Data Points =  %d   Num. of Data Undefined = 0\n", sec5_info->num_data_points);
